refactor(core/cipher): move `cipher` out of `auth`
