+++ core/cipher/encryptors/mock_encryptor.go
