+++ core/cipher/encryptors/wechat_pay_encryptor.go
