+++ core/cipher/encryptors/wechat_pay_encryptor_test.go
